datepicker, javascript for school year

+++ StudInfoSys/Controllers/RegistrationController.cs
@@ -10,6 +10,7 @@ using System.Web.Mvc;
 using StudInfoSys.Models;
 using StudInfoSys.Repository;
 using StudInfoSys.ViewModels;
+using StudInfoSys.Helpers;
 
 namespace StudInfoSys.Controllers
 {
@@ -60,7 +61,6 @@ namespace StudInfoSys.Controllers
         /// <returns></returns>
         public ActionResult Create(int id)
         {
-            //var periods = _unitOfWork.PeriodRepository.GetAll();
             var registrationViewModel = new RegistrationViewModel
                                             {
                                                 StudentId = id,
@@ -74,9 +74,6 @@ namespace StudInfoSys.Controllers
             return View(registrationViewModel);
         }
 
-        //
-        // POST: /Registration/Create
-
         [HttpPost]
         public ActionResult Create(RegistrationViewModel registrationViewModel)
         {
@@ -85,14 +82,6 @@ namespace StudInfoSys.Controllers
                 var registration = MapRegistrationViewModelToRegistration(registrationViewModel);
                 _unitOfWork.RegistrationRepository.Insert(registration);
                 _unitOfWork.RegistrationRepository.Save();
-                //try
-                //{
-                //    _unitOfWork.RegistrationRepository.Save();
-                //}
-                //catch (DbEntityValidationException ex)
-                //{
-                //    Debug.WriteLine("========================================\nDEBUG: " + ex.ToString());
-                //}
                 
                 return RedirectToAction("Index", new { id = registrationViewModel.StudentId });
             }
@@ -101,9 +90,6 @@ namespace StudInfoSys.Controllers
             registrationViewModel.DegreesList = new SelectList(_unitOfWork.DegreeRepository.GetAll(), "Id", "Title", registrationViewModel.DegreeId);
             return View(registrationViewModel);
         }
-
-        //
-        // GET: /Registration/Edit/5
 
         public ActionResult Edit(int id = 0)
         {
@@ -119,16 +105,12 @@ namespace StudInfoSys.Controllers
             return View(registrationViewModel);
         }
 
-        //
-        // POST: /Registration/Edit/5
-
         [HttpPost]
         public ActionResult Edit(RegistrationViewModel registrationViewModel)
         {
             if (ModelState.IsValid)
             {
                 var registration = MapRegistrationViewModelToRegistration(registrationViewModel);
-                //_unitOfWork.RegistrationRepository.GetAll();
                 _unitOfWork.RegistrationRepository.Update(registration);
                 _unitOfWork.RegistrationRepository.Save();
                 return RedirectToAction("Index", new { id = registrationViewModel.StudentId });
@@ -138,9 +120,6 @@ namespace StudInfoSys.Controllers
             registrationViewModel.DegreesList = new SelectList(_unitOfWork.SemesterRepository.GetAll(), "Id", "Title", registrationViewModel.DegreeId);
             return View(registrationViewModel);
         }
-
-        //
-        // GET: /Registration/Delete/5
 
         public ActionResult Delete(int id = 0)
         {
@@ -152,26 +131,28 @@ namespace StudInfoSys.Controllers
             return View(registration);
         }
 
-        //
-        // POST: /Registration/Delete/5
-
         [HttpPost, ActionName("Delete")]
         public ActionResult DeleteConfirmed(int id)
         {
             Registration registration = _unitOfWork.RegistrationRepository.GetById(id);
+            if (registration == null)
+            {
+                return HttpNotFound();
+            }
+
+            // If registration record has at least one related grade record, deletion is prohibited
+            if (registration.SubjectGradesRecords.Any(sgr => sgr.IsDeleted == false))
+            {
+                throw new HttpException("You are not allowed to delete this registration record because it has related grade records");
+            }
+
             var studentId = registration.Student.Id;
             _unitOfWork.RegistrationRepository.Delete(registration);
             _unitOfWork.RegistrationRepository.Save();
 
-            //TODO: create a RegistrationViewModel that includes StudentId as one of its properties
             return RedirectToAction("Index", new { id = studentId });
         }
 
-        //protected override void Dispose(bool disposing)
-        //{
-        //    db.Dispose();
-        //    base.Dispose(disposing);
-        //}
 
 
         private Registration MapRegistrationViewModelToRegistration(RegistrationViewModel registrationsViewModel)
@@ -203,12 +184,12 @@ namespace StudInfoSys.Controllers
                            StudentId = registrations.Student.Id
                        };
         }
+        
+        protected override void OnException(ExceptionContext filterContext)
+        {
+            //Log error
+            Log.WriteLog(Properties.Settings.Default.LogErrorFile, filterContext.Exception.ToString());
+        }
+
     }
 }
-
-//Id = registrationsViewModel.Id,
-//                           DateOfRegistration = registrationsViewModel.DateOfRegistration,
-//                           Degree = _unitOfWork.DegreeRepository.GetById(registrationsViewModel.DegreeId),
-//                           SchoolYearFrom = registrationsViewModel.SchoolYearFrom,
-//                           SchoolYearTo = registrationsViewModel.SchoolYearTo,
-//                           Semester = _unitOfWork.SemesterRepository.GetById(registrationsViewModel.SemesterId)
